deleted TextChat
